correct "delay" parameter in original algorithm
Per the paper on which this library's algorithm is based:

"The delay in the second then clause must be long enough so that, if
another process j read y equal to zero in the first if statement before i
set y equal to i, then j will either enter the second then clause or else
execute the critical section and reset y to zero before i finishes executing
the delay statement. (This delay is allowed because it is executed only if
there is contention.) It is shown in Section 3 that this algorithm guarantees
mutual exclusion and is deadlock free. However, an individual process may
be starved."

http://research.microsoft.com/en-us/um/people/lamport/pubs/fast-mutex.pdf